Create flash-messages.component.js
I'm new at this but I was trying to help remove a bug well in regards to this working with Foundation 6. I couldn't find this file in the github, but it resides in the lib-dist folder when using Angular 2. Anyways, the "alert" in the class was triggering the foundation alert css and even when you added class "success" to the flash message it was being overridden by that default class. When removing "alert" from the default class it worked as expected. Again sorry I'm new at this.
line 34 I removed "alert" from the div class for the template.